pyproject.toml
[project]
name = "yolo-trainer"
version = "0.1.0"
description = "Desktop tool scaffold for preparing and training YOLO datasets from STEM ZC images."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "pillow>=10.0",
    "PySide6>=6.7",
    "tifffile>=2024.0",
]

[project.scripts]
yolo-trainer = "yolo_trainer.app:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
